Update deployment files to fix #43 and to fix #46.
The common package is now baked into all containers directly instead of relying on a separate base image.
The docker-compose.yml file controls the user that each container is run as.
The default docker-compose.yml now starts all Classy services (including autotest and the grader service). It can be extended as is done in docker-compose.310.yml or overridden as in docker-compose.340.yml.

Notes:
- The .env file is currently baked into the container; probably want to avoid this. See #79.
- For the Portal service, git should now be properly configured during the container build; no need to configure it programmatically.